ui: Clear photos when navigating between albums
When the user navigates from Album A to Album B, Album A's photos
were kept in the store, because they're on the same route and
subsequently use the same component instance. This causes the photos
in Album A to appear briefly on Album B before the request for Album B's
photos is complete. Clearing the photos before sending the request
fixes this.